+++ src/parse_method.rs
@@ -4,6 +4,7 @@ use syn::spanned::Spanned;
 pub struct Method {
     visibility: syn::Visibility,
     name: syn::Ident,
+    generics: syn::Generics,
     borrow: syn::Ident,
     tail_args: syn::punctuated::Punctuated<syn::FnArg, syn::Token![,]>,
     tail_args_names: Vec<syn::Ident>,
@@ -15,6 +16,7 @@ impl Parse for Method {
         let visibility = input.parse()?;
         let _: syn::Token![fn] = input.parse()?;
         let name: syn::Ident = input.parse()?;
+        let generics = input.parse()?;
         let args_input;
         let _ = syn::parenthesized!(args_input in input);
         let borrow = syn::Ident::new(match args_input.parse()? {
@@ -50,6 +52,7 @@ impl Parse for Method {
         Ok(Method {
             visibility,
             name,
+            generics,
             tail_args,
             tail_args_names,
             borrow,
@@ -63,6 +66,7 @@ impl quote::ToTokens for Method {
         let Method {
             visibility,
             name,
+            generics,
             borrow,
             tail_args,
             tail_args_names,
@@ -71,7 +75,7 @@ impl quote::ToTokens for Method {
         } = self;
 
         let result = quote::quote! {
-            #visibility fn #name(&self, #tail_args) -> #return_ty {
+            #visibility fn #name #generics(&self, #tail_args) -> #return_ty {
                 //   v-v-- need to be apart to avoid "unexpected token"
                 self.0 .#borrow().#name(#(#tail_args_names)*)
             }